backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlmodel==0.0.21
langdetect==1.0.9
python-dotenv==1.0.1
httpx==0.27.0
websockets==12.0
jinja2==3.1.4
itsdangerous==2.2.0
starlette==0.37.2
orjson==3.10.7
